+++ api/application/controllers/Invoice_master.php
@@ -1,26 +1,34 @@
 <?php
 defined('BASEPATH') or exit('No direct script access allowed');
-require_once FCPATH. "/vendor/autoload.php";
+require_once FCPATH . "/vendor/autoload.php";
 
-class Invoice_master extends CI_Controller{
-    public function __construct() {
+class Invoice_master extends CI_Controller
+{
+
+   
+    private $check_permission;
+    public function __construct()
+    {
         parent::__construct();
-        $this->jwt_token->get_verified_token();
         $this->load->model('invoice_model');
-         $this->load->helper('download');
-       
+        $this->load->helper('download');
+        $this->check_permission = $this->fx->check_user_permission(5);
     }
 
 
     function insert_invoice_data()
     {
+        if ($this->check_permission[0]['add_records'] == 0) {
+            echo "you have not access this page";
+            return;
+        }
 
         $this->form_validation->set_rules('invoice_no', 'Invoice No', 'required|is_unique[invoice_master.invoice_no]');
         if (!$this->form_validation->run()) {
             $error = $this->form_validation->error_array();
 
             echo json_encode(['errors' => $error]);
-           return;
+            return;
         }
 
         $form_data = $this->input->post();
@@ -40,7 +48,7 @@ class Invoice_master extends CI_Controller{
     }
 
 
-   
+
     function client_autocomplete()
     {
         $name = $this->input->post();
@@ -63,39 +71,29 @@ class Invoice_master extends CI_Controller{
         $data = json_decode(file_get_contents('php://input'), true);
         $paggination_records =  $this->fx->paggination_data($data);
         $responce = $this->invoice_model->records($paggination_records);
-        $this->fx->Responce(200,true,$responce);
+        $this->fx->Responce(200, true, $responce);
         return;
     }
 
-    
+
 
     function invoice_pdf()
     {
         if (isset($_GET['id'])) {
             $id = $_GET['id'];
-          
+
             $data['invoice_details'] = $this->invoice_model->get_invoice_pdfData($id);
         }
         $mr = '3px';
         $mpdf = new \Mpdf\Mpdf();
         $mpdf->showImageErrors = true;
         ob_start();
-
-       
-       
         $html = $this->load->view('pdf_html', $data, true);
-
-
         $mpdf->WriteHTML($html);
-
         $data = json_decode($data['invoice_details']);
         $clients = $data->client_details;
-
         $file = 'invoice_files/' . $clients[0]->invoice_no . '.pdf';
-
         $mpdf->Output($file, 'I');
-
-        
     }
 
 
@@ -105,10 +103,10 @@ class Invoice_master extends CI_Controller{
     {
 
         $id = json_decode(file_get_contents('php://input'), true);
-            
-           
+
+
         $data['invoice_details'] = $this->invoice_model->get_invoice_pdfData($id);
-        
+
         $mr = '3px';
         $mpdf = new \Mpdf\Mpdf();
         $mpdf->showImageErrors = true;
@@ -131,6 +129,10 @@ class Invoice_master extends CI_Controller{
 
     function update_invoice_data()
     {
+        if ($this->check_permission[0]['update_records'] == 0) {
+            echo "you have not access this page";
+            return;
+        }
         $form_data = $this->input->post();
 
         $responce =  $this->invoice_model->update_data($form_data);
@@ -145,6 +147,10 @@ class Invoice_master extends CI_Controller{
     function delete_invoice_data()
     {
 
+        if ($this->check_permission[0]['delete_records'] == 0) {
+            echo "you have not access this page";
+            return;
+        }
         $id = json_decode(file_get_contents('php://input'), true);
         $responce = $this->invoice_model->delete_data($id);
         echo $responce;
@@ -154,16 +160,14 @@ class Invoice_master extends CI_Controller{
 
     function edit_invoice_data()
     {
-
         $id = json_decode(file_get_contents('php://input'), true);
-       
         $responce = $this->invoice_model->edit_data($id);
-
         echo  $responce;
         return;
-       
-
     }
 
+    function get_permission(){
+        $responce= $this->invoice_model->permission_details();
+        $this->fx->Responce(200,true,$responce);
+       }
 }
-
